added pages to features refactor
