+++ web/src/App.tsx
@@ -1,15 +1,15 @@
 import { useEffect, useState } from 'react'
 import { Routes, Route, Navigate, useNavigate } from 'react-router-dom'
 import AppLayout from './components/AppLayout'
-import Home from './pages/Home'
-import Library from './pages/Library'
-import MangaDetail from './pages/MangaDetail'
-import Reader from './pages/Reader'
-import Settings from './pages/Settings'
-import Discover from './pages/Discover'
-import Queue from './pages/Queue'
-import Login from './pages/Login'
-import Setup from './pages/Setup'
+import Home from './features/home'
+import Library from './features/library'
+import MangaDetail from './features/manga-detail'
+import Reader from './features/reader'
+import Settings from './features/settings'
+import Discover from './features/discover'
+import Queue from './features/queue'
+import Login from './features/login'
+import Setup from './features/setup'
 import { api, getToken, clearToken } from './api'
 import { ROUTES } from './lib/routes'
