+++ web/src/components/SettingRow.tsx
@@ -1,0 +1,11 @@
+export function SettingRow({ label, hint, children }: { label: string; hint: string; children: React.ReactNode }) {
+  return (
+    <div className="flex items-center justify-between gap-4">
+      <div className="min-w-0">
+        <p className="text-sm font-medium">{label}</p>
+        <p className="text-xs text-muted-foreground">{hint}</p>
+      </div>
+      <div className="shrink-0">{children}</div>
+    </div>
+  )
+}
