+++ web/src/features/discover/components/SearchSkeleton.tsx
@@ -1,0 +1,19 @@
+import { Skeleton } from '@/components/ui/skeleton'
+
+export function SearchSkeleton() {
+  return (
+    <div className="space-y-3">
+      {Array.from({ length: 5 }).map((_, i) => (
+        <div key={i} className="flex gap-3 rounded-lg border border-border bg-card p-3">
+          <Skeleton className="w-14 shrink-0 rounded aspect-[2/3]" />
+          <div className="flex-1 space-y-2">
+            <Skeleton className="h-4 w-3/4" />
+            <Skeleton className="h-3 w-1/4" />
+            <Skeleton className="h-3 w-full" />
+            <Skeleton className="h-3 w-5/6" />
+          </div>
+        </div>
+      ))}
+    </div>
+  )
+}
