+++ web/src/features/manga-detail/index.tsx
@@ -1,22 +1,19 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { useParams, useNavigate } from 'react-router-dom'
 import {
-  Download, RefreshCw, HardDrive,
-  CheckCircle2, BookOpen, Clock, Loader2, AlertCircle,
-  Play, ChevronDown, X, ArrowUp, ArrowDown, Trash2,
+  Download, RefreshCw, Loader2, X, ArrowUp, ArrowDown, Trash2, ChevronDown,
 } from 'lucide-react'
-import { api, isAdmin, type QueueItem } from '@/api'
-import type { Chapter, ReadProgress } from '@/types'
+import { api, isAdmin } from '@/api'
 import { Button } from '@/components/ui/button'
-import { Skeleton } from '@/components/ui/skeleton'
-import { cn } from '@/lib/utils'
 import { ContentTypePill } from '@/components/ContentTypePill'
 import { ROUTES } from '@/lib/routes'
-import { useMangaDetail, CHAPTERS_PREVIEW, type FilterMode } from '@/features/manga-detail/hooks/useMangaDetail'
-
-function formatDate(iso: string) {
-  return new Date(iso).toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
-}
+import { cn } from '@/lib/utils'
+import { useMangaDetail, CHAPTERS_PREVIEW, type FilterMode } from './hooks/useMangaDetail'
+import { ChapterRow } from './components/ChapterRow'
+import {
+  SectionHeading, CoverImg, ChapterListSkeleton,
+  ReaderModeCard, AutoDownloadCard, ExplicitCard, DownloadDirCard,
+} from './components/SidebarCards'
 
 export default function MangaDetail() {
   const { id } = useParams<{ id: string }>()
@@ -87,7 +84,7 @@ export default function MangaDetail() {
                   >
                     {h.pendingReadId === h.resumeChapter.id && !h.resumeChapter.downloaded
                       ? <Loader2 className="w-3.5 h-3.5 animate-spin" />
-                      : <Play className="w-3.5 h-3.5 fill-current" />}
+                      : <Download className="w-3.5 h-3.5 fill-current" />}
                     {h.readCount > 0 ? `Resume Ch. ${h.resumeChapter.number}` : `Start Ch. ${h.resumeChapter.number}`}
                   </Button>
                 )}
@@ -309,7 +306,7 @@ export default function MangaDetail() {
                   <div className="p-3">
                     <p className="text-[10px] text-muted-foreground uppercase tracking-wider">Queued</p>
                     <p className="text-sm font-semibold mt-0.5 text-muted-foreground flex items-center gap-1">
-                      <Clock className="w-3 h-3 shrink-0" />{h.pendingCount} ch
+                      <Download className="w-3 h-3 shrink-0" />{h.pendingCount} ch
                     </p>
                   </div>
                 </>}
@@ -336,368 +333,6 @@ export default function MangaDetail() {
           </div>
         </div>
       </div>
-    </div>
-  )
-}
-
-// ——— Chapter row ———
-
-function ChapterRow({
-  chapter, progress, queueItem, pendingRead, onOpen, onCancelDownload,
-}: {
-  chapter: Chapter
-  progress: ReadProgress | null
-  queueItem: QueueItem | null
-  pendingRead: boolean
-  onOpen: () => void
-  onCancelDownload: (queueId: string) => void
-}) {
-  const isCompleted  = progress?.completed === true
-  const isStarted    = progress != null && !isCompleted
-  const pct          = chapter.page_count > 0 && progress
-    ? Math.round((progress.current_page / chapter.page_count) * 100)
-    : 0
-  const isDownloaded = chapter.downloaded || queueItem?.status === 'done'
-  const isPending    = !isDownloaded && queueItem?.status === 'pending'
-  const isActive     = !isDownloaded && queueItem?.status === 'downloading'
-  const isError      = !isDownloaded && queueItem?.status === 'error'
-  const isWaiting    = pendingRead && !isDownloaded
-  const isClickable  = isDownloaded || (chapter.source_id != null && !isError)
-
-  return (
-    <div
-      className={cn(
-        'flex items-center gap-3 px-3 py-2.5 rounded-lg border border-border bg-card transition-colors',
-        isClickable && 'cursor-pointer hover:bg-accent/40',
-        isCompleted && 'opacity-50',
-      )}
-      onClick={isClickable ? onOpen : undefined}
-    >
-      <span className={cn(
-        'w-11 text-center text-xs font-bold py-1.5 rounded-md shrink-0 leading-none',
-        isDownloaded ? 'bg-primary/20 text-primary' : 'bg-muted text-muted-foreground',
-      )}>
-        {chapter.number}
-      </span>
-
-      <div className="flex-1 min-w-0">
-        <p className="text-sm font-medium truncate">
-          {chapter.title ?? `Chapter ${chapter.number}`}
-        </p>
-        <div className="flex items-center gap-1.5 text-[11px] text-muted-foreground mt-0.5">
-          <span>{formatDate(chapter.created_at)}</span>
-          {chapter.page_count > 0 && (
-            <>
-              <span className="opacity-40">•</span>
-              <span>{chapter.page_count} Pages</span>
-            </>
-          )}
-          {isStarted && (
-            <>
-              <span className="opacity-40">•</span>
-              <span className="text-primary">{progress!.current_page + 1}/{chapter.page_count}</span>
-            </>
-          )}
-          {isWaiting && (
-            <>
-              <span className="opacity-40">•</span>
-              <span className="text-primary">Opening when ready…</span>
-            </>
-          )}
-        </div>
-        {isStarted && (
-          <div className="w-full h-0.5 rounded-full bg-muted mt-1.5">
-            <div className="h-full rounded-full bg-primary transition-all" style={{ width: `${pct}%` }} />
-          </div>
-        )}
-      </div>
-
-      <div className="flex items-center gap-1 shrink-0" onClick={(e) => e.stopPropagation()}>
-        {isCompleted && <CheckCircle2 className="w-3.5 h-3.5 text-muted-foreground mr-0.5" />}
-        {isDownloaded ? (
-          <button className="w-7 h-7 flex items-center justify-center rounded-md hover:bg-accent transition-colors" onClick={onOpen} title="Read">
-            <BookOpen className="w-3.5 h-3.5 text-primary" />
-          </button>
-        ) : isActive || isWaiting ? (
-          <Loader2 className="w-4 h-4 text-primary animate-spin" />
-        ) : isPending ? (
-          <button
-            onClick={() => onCancelDownload(queueItem!.id)}
-            className="flex items-center gap-1 px-2 py-1 rounded-md text-[11px] font-medium bg-muted text-muted-foreground hover:bg-destructive/10 hover:text-destructive transition-colors"
-          >
-            <Clock className="w-3 h-3 shrink-0" />
-            Queued
-            <X className="w-3 h-3 shrink-0" />
-          </button>
-        ) : isError ? (
-          <button
-            className="w-7 h-7 flex items-center justify-center rounded-md hover:bg-accent transition-colors text-destructive"
-            onClick={onOpen}
-            title="Retry"
-          >
-            <AlertCircle className="w-3.5 h-3.5" />
-          </button>
-        ) : chapter.source_id ? (
-          <button
-            className="w-7 h-7 flex items-center justify-center rounded-md hover:bg-accent transition-colors"
-            onClick={onOpen}
-            title="Download & read"
-          >
-            <Download className="w-3.5 h-3.5 text-muted-foreground" />
-          </button>
-        ) : (
-          <HardDrive className="w-3.5 h-3.5 text-muted-foreground/40" />
-        )}
-      </div>
-    </div>
-  )
-}
-
-// ——— Sidebar setting cards ———
-
-function ReaderModeCard({ mangaId, value }: { mangaId: string; value: string | null }) {
-  const [current, setCurrent] = useState(value)
-  const options: { label: string; v: string | null }[] = [
-    { label: 'Global', v: null },
-    { label: 'Paged',  v: 'paged' },
-    { label: 'Scroll', v: 'scroll' },
-  ]
-
-  async function handleSelect(i: number) {
-    const v = options[i].v
-    setCurrent(v)
-    await api.setMangaReaderMode(mangaId, v).catch(() => {})
-  }
-
-  return (
-    <SettingSegmentCard
-      label="Reader Mode"
-      description={
-        current == null
-          ? 'Uses the global reader mode setting.'
-          : current === 'paged'
-          ? 'One page at a time, tap/click to advance.'
-          : 'All pages in a continuous vertical scroll.'
-      }
-    >
-      <SegmentControl
-        options={options.map((o) => o.label)}
-        active={options.findIndex((o) => o.v === current)}
-        onSelect={handleSelect}
-      />
-    </SettingSegmentCard>
-  )
-}
-
-function AutoDownloadCard({ mangaId, value }: { mangaId: string; value: boolean | null }) {
-  const [current, setCurrent] = useState(value)
-  const options: { label: string; v: boolean | null }[] = [
-    { label: 'Global', v: null },
-    { label: 'Always', v: true },
-    { label: 'Never',  v: false },
-  ]
-
-  async function handleSelect(i: number) {
-    const v = options[i].v
-    setCurrent(v)
-    await api.setMangaAutoDownload(mangaId, v).catch(() => {})
-  }
-
-  return (
-    <SettingSegmentCard
-      label="Auto-download"
-      description={
-        current === null
-          ? 'Follows the global auto-download setting.'
-          : current
-          ? 'New chapters download automatically.'
-          : 'New chapters are never auto-downloaded.'
-      }
-    >
-      <SegmentControl
-        options={options.map((o) => o.label)}
-        active={options.findIndex((o) => o.v === current)}
-        onSelect={handleSelect}
-      />
-    </SettingSegmentCard>
-  )
-}
-
-function ExplicitCard({ mangaId, value }: { mangaId: string; value: boolean }) {
-  const [current, setCurrent] = useState(value)
-
-  async function toggle() {
-    const next = !current
-    setCurrent(next)
-    await api.setMangaExplicit(mangaId, next).catch(() => setCurrent(current))
-  }
-
-  return (
-    <div className="rounded-lg bg-card border border-border p-4 space-y-2">
-      <div className="flex items-center justify-between gap-3">
-        <p className="text-[10px] font-semibold uppercase tracking-wider text-muted-foreground">
-          Explicit Content
-        </p>
-        <button
-          type="button"
-          onClick={toggle}
-          className={`relative w-10 h-[22px] rounded-full transition-colors ${current ? 'bg-orange-500' : 'bg-muted'}`}
-        >
-          <span className={`absolute top-[3px] w-4 h-4 rounded-full bg-white shadow transition-transform ${current ? 'translate-x-5' : 'translate-x-[3px]'}`} />
-        </button>
-      </div>
-      <p className="text-xs text-muted-foreground">
-        {current ? 'Marked 18+ — only users with explicit access can see this.' : 'Not marked explicit — visible to all users.'}
-      </p>
-    </div>
-  )
-}
-
-function DownloadDirCard({ mangaId, value }: { mangaId: string; value: string | null }) {
-  const [draft, setDraft] = useState(value ?? '')
-  const [saving, setSaving] = useState(false)
-
-  async function handleSave(v: string | null) {
-    setSaving(true)
-    await api.setMangaDownloadDir(mangaId, v).catch(() => {})
-    setSaving(false)
-  }
-
-  const isDirty = draft !== (value ?? '')
-  const isDefault = !value
-
-  return (
-    <div className="rounded-lg bg-card border border-border p-4 space-y-3">
-      <p className="text-[10px] font-semibold uppercase tracking-wider text-muted-foreground">
-        Download Path
-      </p>
-      <div className="flex gap-2">
-        <input
-          type="text"
-          value={draft}
-          onChange={(e) => setDraft(e.target.value)}
-          onKeyDown={(e) => { if (e.key === 'Enter') handleSave(draft.trim() || null) }}
-          placeholder="/path/to/manga"
-          className="flex-1 bg-muted border border-transparent focus:border-ring rounded-md px-3 py-1.5 text-xs outline-none transition-colors"
-        />
-        {isDirty && (
-          <button
-            onClick={() => handleSave(draft.trim() || null)}
-            disabled={saving}
-            className="px-3 py-1.5 rounded-md bg-primary text-primary-foreground text-xs font-semibold hover:opacity-90 transition-opacity"
-          >
-            Save
-          </button>
-        )}
-        {!isDirty && !isDefault && (
-          <button
-            onClick={() => { setDraft(''); handleSave(null) }}
-            className="px-3 py-1.5 rounded-md bg-muted border border-border text-xs text-muted-foreground hover:text-foreground transition-colors"
-          >
-            Reset
-          </button>
-        )}
-      </div>
-      <p className="text-[10px] text-muted-foreground leading-relaxed">
-        {isDefault
-          ? 'Uses default: _downloads/{title}/ inside your manga dir.'
-          : `Chapters save to ${value}`}
-      </p>
-    </div>
-  )
-}
-
-// ——— Reusable primitives ———
-
-function SettingSegmentCard({
-  label,
-  description,
-  children,
-}: {
-  label: string
-  description: string
-  children: React.ReactNode
-}) {
-  return (
-    <div className="rounded-lg bg-card border border-border p-4 space-y-3">
-      <p className="text-[10px] font-semibold uppercase tracking-wider text-muted-foreground">{label}</p>
-      {children}
-      <p className="text-[10px] text-muted-foreground leading-relaxed">{description}</p>
-    </div>
-  )
-}
-
-function SegmentControl({
-  options,
-  active,
-  onSelect,
-}: {
-  options: string[]
-  active: number
-  onSelect: (i: number) => void
-}) {
-  return (
-    <div className="flex rounded-lg bg-muted p-0.5 gap-0.5">
-      {options.map((label, i) => (
-        <button
-          key={label}
-          onClick={() => onSelect(i)}
-          className={cn(
-            'flex-1 text-[11px] font-medium py-1.5 rounded-md transition-colors',
-            active === i
-              ? 'bg-card text-foreground shadow-sm'
-              : 'text-muted-foreground hover:text-foreground',
-          )}
-        >
-          {label}
-        </button>
-      ))}
-    </div>
-  )
-}
-
-function SectionHeading({ children }: { children: React.ReactNode }) {
-  return (
-    <h2 className="text-xs font-semibold uppercase tracking-wider text-muted-foreground flex items-center gap-2">
-      <span className="w-0.5 h-3.5 bg-primary rounded-full inline-block shrink-0" />
-      {children}
-    </h2>
-  )
-}
-
-function CoverImg({ coverUrl, mangaId }: { coverUrl: string | null; mangaId: string }) {
-  const [failed, setFailed] = useState(false)
-  const src = !failed && coverUrl?.startsWith('http') ? coverUrl : api.coverUrl(mangaId)
-  if (failed) {
-    return (
-      <div className="w-36 shrink-0 rounded-xl aspect-[2/3] bg-muted/60 flex items-center justify-center text-4xl shadow-2xl ring-1 ring-white/10">
-        📖
-      </div>
-    )
-  }
-  return (
-    <img
-      src={src}
-      alt=""
-      className="w-36 shrink-0 rounded-xl aspect-[2/3] object-cover bg-muted shadow-2xl ring-1 ring-white/10"
-      onError={() => setFailed(true)}
-    />
-  )
-}
-
-function ChapterListSkeleton() {
-  return (
-    <div className="space-y-1.5">
-      {Array.from({ length: 5 }).map((_, i) => (
-        <div key={i} className="flex items-center gap-3 px-3 py-2.5 rounded-lg border border-border bg-card">
-          <Skeleton className="w-11 h-7 rounded-md shrink-0" />
-          <div className="flex-1 space-y-1.5">
-            <Skeleton className="h-4 w-1/2" />
-            <Skeleton className="h-3 w-1/3" />
-          </div>
-          <Skeleton className="w-7 h-7 rounded-md shrink-0" />
-        </div>
-      ))}
     </div>
   )
 }
