--- web/src/pages/Login.tsx
@@ -1,77 +1,0 @@
-import { useState } from 'react'
-import { useNavigate } from 'react-router-dom'
-import { Loader2 } from 'lucide-react'
-import { api, setToken } from '@/api'
-import { ROUTES } from '@/lib/routes'
-import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-
-export default function Login() {
-  const navigate = useNavigate()
-  const [username, setUsername] = useState('')
-  const [password, setPassword] = useState('')
-  const [error, setError] = useState('')
-  const [loading, setLoading] = useState(false)
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault()
-    setError('')
-    setLoading(true)
-    try {
-      const res = await api.login(username.trim(), password)
-      setToken(res.token, res.username, res.role, res.allow_explicit)
-      navigate(ROUTES.home, { replace: true })
-    } catch {
-      setError('Invalid username or password.')
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  return (
-    <div className="min-h-screen bg-background flex items-center justify-center p-4">
-      <div className="w-full max-w-sm space-y-8">
-        {/* Brand */}
-        <div className="text-center space-y-1">
-          <h1 className="text-4xl font-black tracking-tight text-primary">*ARRgh</h1>
-          <p className="text-sm text-muted-foreground">Sign in to your library</p>
-        </div>
-
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div className="space-y-2">
-            <label className="text-sm font-medium text-foreground">Username</label>
-            <Input
-              type="text"
-              autoComplete="username"
-              value={username}
-              onChange={(e) => setUsername(e.target.value)}
-              placeholder="username"
-              required
-              autoFocus
-            />
-          </div>
-          <div className="space-y-2">
-            <label className="text-sm font-medium text-foreground">Password</label>
-            <Input
-              type="password"
-              autoComplete="current-password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              placeholder="••••••••"
-              required
-            />
-          </div>
-
-          {error && (
-            <p className="text-sm text-destructive">{error}</p>
-          )}
-
-          <Button type="submit" className="w-full" disabled={loading}>
-            {loading ? <Loader2 className="w-4 h-4 animate-spin mr-2" /> : null}
-            Sign In
-          </Button>
-        </form>
-      </div>
-    </div>
-  )
-}
